SaveLogInfo is now commented out due to a cluster issue.
For some reason unknown, matlab on the cluster cannot locate the
standard UNIX gitpath (`/usr/bin`) and fails to get version
information. I can easily access said path and the git file needed is
there, it just returns false if I use the `exist` command in matlab.
Git commands also function normally on the cluster.
This  did not use to happen before TbTb on the cluster.